casos de uso ateh 28 jan 24 (#23)
* config: habilitando coertura de codigo

* config: habilitando coertura de codigo

* config: pasta de workflow dentro de KnnSystems

* config: actions soh no ramo integrador de casos de uso temporariamente

* ucs01: autenticar usuario (#3)

* refact: corrigindo nome de package servic para service

* refact: correcao padrao incorreto constantes dos enums

* feat: primeira impl funcional para ucs01

* feat: senha provisoria

* tests: correcao em atualizacao de senha

* refact: corrigindo hierarquia usuario pessoa

* feat: adicionando requisito de usuario ativo

* config: diretorio de workflows alterado para execucao de teste de ci

* ucs02: manter usuario (#4)

* refact: corrigindo nome de package servic para service

* refact: correcao padrao incorreto constantes dos enums

* feat: primeira impl funcional para ucs01

* feat: senha provisoria

* tests: correcao em atualizacao de senha

* refact: corrigindo hierarquia usuario pessoa

* feat: perfis para aplicacao

* feat: apenas administradores podem manter outros usuarios

* feat: endpoint para registro usuario

* doc: adicionando referencia ao uso do github actions no projeto

* config: reabilitando ci nas linhas principais

* refact: corrigindo hierarquia pacote infrastructure

* feat: impl para fluxo alternativo 1.1 do caso de uso

* refact: controller proprio para o caso de uso 02

* refact: passando para service logica de cadastro

* feat: impl para entidade telefone

* mend

* docs: adicionando script de criacao base no versionamento

* feat: impl para fluxo alternativo 1 do caso de uso

* feat: impl para fluxo alternativo 4 do caso de uso

* feat: impl para fluxo principal do caso de uso

* feat: impl para fluxo alternativo 2.1

* feat: implementando restricao ao papel de administrador para endpoint do manter usuario

* tests: testes que secretaria nao acessa endpoint do caso de uso

* feat: ativacao de usuario inativo

* feat: ativacao e inativacao de usuario

* ucs15: relatorio apartamentos (#5)

* refact: perfil de sindico para extracao de relatorios

* feat: impl para fluxo principal de apartamentos

* tests: teste para relatorio com unico apartamento

* test: validando resposta completa

* tests: fluxo principal do caso de uso

* config: habilitando flyway no projeto (#6)

* config: habilitando flyway no projeto

* config: valores default para github actions runner

* refact: encapsulando testes em transacoes para garantir sua independencia

* ucs10: manter apartamento (#7)

* feat: fluxo alternativo 1.1 do caso de uso

* tests: excecoes do fluxo alternativo 1

* feat: impl para fluxo de inclusao de apartamento

* feat: ativar apartamento ao criar

* feat: consulta de apartamentos

* feat: inativacao de apartamento

* feat: adicionando restricao de perfil no caso de uso

* ucs17: manter morador (#8)

* feat: fluxo de excecao na inclusao

* tests: mais cenarios de excecao

* feat: cadastro de morador

* feat: consulta de morador por nome ou cpf

* ucs03: gerar extrato financeiro (#9)

* feat: classe de transacao financeira

* tests: valor zero para transacao

* feat: instanciacao extrato financeiro

* feat: inclui data de emissão

* feat: inclui calculo de total de entradas e saldo

* feat: atualizada total de saidas para transacao de despesa

* feat: validacao de transacao estah no periodo do extrato

* feat: esboco de service para gerar o extrato

* feat: impl mocada para servicoes de extrato financeiro

* feat: tratando consulta vazia

* feat: tratamento de erro para params invalidos de consulta

* config: corsconfiguration para requests do frontend (#10)

* config: readme do repositorio e h2 como dep para testes (#1) (#11)

* config: adicionando dep h2 para testes unitarios

* docs: adicionando readme.md ao projeto

* doc: descricao de metodologia workflow

* feat: fluxo alternativo 2 do caso de uso (#13)

* ucs01: validacao senha e cpf na gestao de usuarios (#14)

* feat: trava cadastro de usuario com menos de 8 posicoes

* feat: trava para cadastro de senha atender ao documento de especificacao

* refact: renomeando controller de usuario para padrao usado em outros

* feat: validacoes de cpf e senha nos dtos de autenticacao

* ucs09: principais fluxos caso de uso fornecedor (#15)

* feat: impl parcial para fornecedor - depende de contrato

* refact: ajustes em entities e scripts sql

* feat: fluxo alternativo 1.3 do caso de uso

* feat: adiciona atributo telefone a responsavel para implementar caso de uso

* test: testes com telefone para responsavel

* feat: colocando obrigatoriedade nos campos

* tests: fluxo alternativo 1.2

* feat: fluxo alternativo 1

* feat: esboco de impl para consulta fornecedores

* test: teste para fluxo alternativo 4

* feat: fluxo principal de consulta

* feat: adicionando facade com impl stub descrito na documentacao

* feat: inativacao de fornecedor

* ucs11 e ucs12 (#17)

* feat: impl inicial para o fluxo alternativo 1.1 do caso de uso

* tests: teste para fluxo alternativo 1.2

* feat: fluxo alternativo 1.3

* feat: fluxo alternativo 1 do caso de uso

* feat: impl basica para fluxo principal

* feat: fluxo principal do caso de uso

* feat: inativar contrato

* feat: esqueleto para endpoint de reajuste

* feat: dto com parametros de reajuste

* feat: fluxo alternativo 2.2.1 do caso de uso

* feat: fluxo alternativo 2 para API

* feat: ajustes no calculo de reajuste

* feat: avanco em cenarios de excecao do ucs12

* feat: cenario de fluxo feliz do ucs12

* feat: completando ucs12

* ucs05: manter pagamento (#20)

* feat: terminando impl de rescisao do ucs11

* feat: impl para fluxo alternativo 1.5

* feat: fluxo alternativo 1.4

* feat; fluxo allternativo 1.3

* feat: calculadora dias uteis e validacao numero contrato

* feat: fluxo alternativo 1.6

* feat: fluxo alternativo 1.7

* feat; validacao data no futuro

* feat: fluxo alternativo 1

* feat: fluxo alternativo 4 e principal

* feat: fluxo inativar

* ucs16: exclusao (#21)

* feat: exclusao de fornecedor

* feat: fluxo alternativo 1 do caso de uso

* feat: fluxo alternativo 4 do caso de uso

* feat: fluxo alternativo 5

* ucs15: mais relatorios (#22)

* feat: fluxo alternativo 2 do caso de uso

* feat: fluxo alternativo 2 do caso de uso

* feat: fluxo alternativo 3 do caso de uso

* feat: fluxo alternativo 6 do caso de uso

+++ KnnSystem/src/main/java/com/knnsystem/api/dto/ApartamentoFormularioDTO.java
@@ -11,7 +11,7 @@ import org.hibernate.validator.constraints.br.CPF;
 
 public record ApartamentoFormularioDTO(
 		@NotNull(message = "número do apartamento deve ser preenchido")
-		int numeroDoApartamento,
+		Integer numeroDoApartamento,
 		@NotBlank(message = "bloco deve ser preenchido")
 		String bloco,
 		@NotBlank(message = "nome do proprietário deve ser preenchido")
@@ -35,9 +35,9 @@ public record ApartamentoFormularioDTO(
 		@Email(message = "e-mail do morador inválido")
 		String emailDoMorador,
 		@NotNull(message = "metragem do imóvel deve ser preenchida")
-		int metragemDoImovel,
+		Integer metragemDoImovel,
 
-		int id) {
+		Long id) {
 
 	public ApartamentoFormularioDTO(Apartamento apartamento){
 		this(apartamento.getNumApt(), apartamento.getBlocoApt(),
